requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.36.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# NLP & Text Processing
sentence-transformers>=2.2.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# API Documentation
flasgger>=0.9.7

# Monitoring & Logging
loguru>=0.7.0
